ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
numpy==1.26.4
faiss-cpu==1.8.0.post1
open_clip_torch==2.26.1
torch==2.3.1
opencv-python==4.8.1.78
scikit-image==0.21.0
scipy==1.11.4
segment-anything==1.0
tqdm==4.66.1
